+++ palmettobug/ImageProcessing/ImageAnalysisClass.py
@@ -927,11 +927,11 @@ class ImageAnalysis:
                     _RUST_OK = False 
                 if _RUST_OK:
                     print('rusty boolin')
-                    output = rsc.mask_boolean_rust(np.ascontiguousarray(mask1, dtype=np.int32), 
-                                                   np.ascontiguousarray(mask2, dtype=np.int32), 
+                    output = rsc.mask_boolean_rust(np.ascontiguousarray(mask1, dtype=np.uint), 
+                                                   np.ascontiguousarray(mask2, dtype=np.uint), 
                                                    kind = kind, 
-                                                   object_threshold = np.int32(object_threshold), 
-                                                   pixel_threshold = np.int32(pixel_threshold), 
+                                                   object_threshold = object_threshold, 
+                                                   pixel_threshold = pixel_threshold, 
                                                    re_order = True)
                 else:
                     output = self._mask_bool(mask1, mask2, kind = kind, object_threshold = object_threshold, pixel_threshold = pixel_threshold)
